Fix bug + move inner class in its own file
