+++ profiler/src/ProfilerEngine/Datadog.Profiler.Native.Linux/CpuProfilerDisableScope.h
@@ -1,0 +1,18 @@
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2022 Datadog, Inc.
+
+#include <cstdint>
+#include <time.h>
+
+struct ManagedThreadInfo;
+
+struct CpuProfilerDisableScope
+{
+public:
+    explicit CpuProfilerDisableScope(ManagedThreadInfo* threadInfo);
+    ~CpuProfilerDisableScope();
+
+private:
+    std::int32_t _timerId;
+    struct itimerspec _oldValue;
+};
